refactor: Move applyMeetingFilters to meeting-utils and move filter to route data loader.
