+++ src/routes/meetings-filtered.tsx
@@ -4,14 +4,18 @@ import { Filter } from "@/components/Filter"
 import { Layout } from "@/components/Layout"
 import { MeetingsSummary } from "@/components/MeetingsSummary"
 import { getNextMeetings } from "@/getMeetings"
-import { buildFilter } from "@/meetings-utils"
+import { applyMeetingFilters, buildFilter } from "@/meetings-utils"
 
 import type { Route } from "./+types/meetings-filtered"
 
 export async function clientLoader({ request }: Route.ClientLoaderArgs) {
   const { searchParams } = new URL(request.url)
-  const meetings = await getNextMeetings(buildFilter(searchParams))
+  const filter = buildFilter(searchParams)
+  const unfilteredMeetings = await getNextMeetings()
+
+  const meetings = applyMeetingFilters(unfilteredMeetings, filter)
   console.log(meetings)
+
   return { meetings }
 }
 
@@ -38,6 +42,7 @@ export default function MeetingsFiltered({ loaderData }: Route.ComponentProps) {
       >
         <MeetingsSummary meetings={loaderData.meetings} />
       </Layout>
+      {/* <Outlet context={loaderData.meetings} /> */}
     </>
   )
 }
